fix(metadata): improve RanobeDB parser robustness and code quality
AI-assisted code review identified and fixed the following issues:

Bug fixes:
- Fix log.error → log.info for success message (was logging at wrong level)
- Add null checks for collections (getTitles, getReleases, getPublishers, getEditions, getTags) to prevent NPE
- Filter null results from stream in parseRanobeDbApiResponse
- Fix exception type: catch DateTimeException instead of DateTimeParseException
- Add Boolean.TRUE.equals() for getOfficial() null safety

Missing functionality:
- Add provider enabled check using appSettingService (consistent with other parsers)

Code cleanup:
- Change getMetadataListByTerm visibility from public to private
- Change Boolean parameters to primitive boolean
- Fix typo: responstObj → responseObj
- Remove unused imports (Instant, DateTimeFormatter, AtomicBoolean)
- Fix type inconsistency: RanobedbSearchResponse.Book.id changed from int to Long
- Extract static TITLE_CASE_PATTERN (avoid recompiling regex on each call)

+++ booklore-api/src/main/java/com/adityachandel/booklore/model/dto/response/ranobedbapi/RanobedbSearchResponse.java
@@ -3,19 +3,18 @@ package com.adityachandel.booklore.model.dto.response.ranobedbapi;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 import java.util.List;
-import java.util.Set;
 
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class RanobedbSearchResponse{
-  private List<Book> books;
-  private String count;
-  private int currentPage;
-  private int totalPages;
+public class RanobedbSearchResponse {
+    private List<Book> books;
+    private String count;
+    private int currentPage;
+    private int totalPages;
 
-  @Data
-  @JsonIgnoreProperties(ignoreUnknown = true)
-  public static class Book {
-    private int id;
-  }
+    @Data
+    @JsonIgnoreProperties(ignoreUnknown = true)
+    public static class Book {
+        private Long id;
+    }
 }
